fix pure helper prettifyNumber
